some changes in the redux-store folder
