+++ package.json
@@ -25,13 +25,14 @@
     "react-modal": "^3.16.1",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.16.0",
+    "redux": "^5.0.1",
     "redux-persist": "^6.0.0",
     "simplebar-react": "^3.2.4",
     "socket.io": "^4.7.2",
     "socket.io-client": "^4.7.2",
     "styled-components": "^6.1.0",
     "sweetalert": "^2.1.2",
-    "vite": "^5.0.12",
+    "vite": "^4.4.0",
     "vite-tsconfig-paths": "^4.3.1",
     "web-vitals": "^2.1.4",
     "yup": "^1.3.2"
