+++ src/redux-store/constants.js
@@ -1,0 +1,10 @@
+const ULRs = {
+  register: 'authentication/register',
+  login: 'authentication/login',
+  logout: 'authentication/logout',
+  currentUser: 'current',
+  updateUser: 'users',
+  country: 'countries',
+};
+
+export default ULRs;
